refactor: remove canonical message layer
