+++ src/tunacode/core/types/tool_registry.py
@@ -5,10 +5,10 @@ from __future__ import annotations
 from collections.abc import Iterable
 from dataclasses import dataclass, field, replace
 from datetime import UTC, datetime
+from enum import Enum
 from typing import Any
 
-from tunacode.types.base import ToolArgs, ToolCallId, ToolName
-from tunacode.types.canonical import CanonicalToolCall, CanonicalToolResult, ToolCallStatus
+from tunacode.types.base import ToolArgs, ToolCallId, ToolName, ToolResult
 
 ERROR_TOOL_CALL_ID_REQUIRED = "tool_call_id is required"
 ERROR_TOOL_CALL_NOT_FOUND = "Tool call not registered: {tool_call_id}"
@@ -17,6 +17,39 @@ TOOL_RECORD_KEY_ARGS = "args"
 TOOL_RECORD_KEY_TOOL = "tool"
 TOOL_RECORD_KEY_TIMESTAMP = "timestamp"
 TOOL_RECORD_KEY_TOOL_CALL_ID = "tool_call_id"
+
+
+class ToolCallStatus(Enum):
+    """Lifecycle status of a tool call."""
+
+    PENDING = "pending"
+    RUNNING = "running"
+    COMPLETED = "completed"
+    FAILED = "failed"
+    CANCELLED = "cancelled"
+
+
+@dataclass(frozen=True, slots=True)
+class ToolCallRecord:
+    """Typed tool call record for runtime lifecycle tracking."""
+
+    tool_call_id: ToolCallId
+    tool_name: ToolName
+    args: ToolArgs
+    status: ToolCallStatus
+    result: ToolResult | None = None
+    error: str | None = None
+    started_at: datetime | None = None
+    completed_at: datetime | None = None
+
+    @property
+    def is_complete(self) -> bool:
+        """Return whether the tool call has finished."""
+        return self.status in (
+            ToolCallStatus.COMPLETED,
+            ToolCallStatus.FAILED,
+            ToolCallStatus.CANCELLED,
+        )
 
 
 def _utc_now() -> datetime:
@@ -33,12 +66,12 @@ def _format_timestamp(timestamp: datetime | None) -> str | None:
 class ToolCallRegistry:
     """Single source of truth for tool call lifecycle."""
 
-    _calls: dict[ToolCallId, CanonicalToolCall] = field(default_factory=dict)
+    _calls: dict[ToolCallId, ToolCallRecord] = field(default_factory=dict)
     _order: list[ToolCallId] = field(default_factory=list)
 
     def register(
         self, tool_call_id: ToolCallId, tool_name: ToolName, args: ToolArgs
-    ) -> CanonicalToolCall:
+    ) -> ToolCallRecord:
         """Register a new tool call."""
         if not tool_call_id:
             raise ValueError(ERROR_TOOL_CALL_ID_REQUIRED)
@@ -48,7 +81,7 @@ class ToolCallRegistry:
             self._calls[tool_call_id] = updated_call
             return updated_call
 
-        call = CanonicalToolCall(
+        call = ToolCallRecord(
             tool_call_id=tool_call_id,
             tool_name=tool_name,
             args=args,
@@ -60,7 +93,7 @@ class ToolCallRegistry:
 
     def start(
         self, tool_call_id: ToolCallId, started_at: datetime | None = None
-    ) -> CanonicalToolCall:
+    ) -> ToolCallRecord:
         """Mark a tool call as running."""
         resolved_started_at = started_at or _utc_now()
         return self._update_call(
@@ -72,9 +105,9 @@ class ToolCallRegistry:
     def complete(
         self,
         tool_call_id: ToolCallId,
-        result: CanonicalToolResult | None = None,
+        result: ToolResult | None = None,
         completed_at: datetime | None = None,
-    ) -> CanonicalToolCall:
+    ) -> ToolCallRecord:
         """Mark a tool call as completed."""
         resolved_completed_at = completed_at or _utc_now()
         return self._update_call(
@@ -88,9 +121,9 @@ class ToolCallRegistry:
         self,
         tool_call_id: ToolCallId,
         error: str | None,
-        result: CanonicalToolResult | None = None,
+        result: ToolResult | None = None,
         completed_at: datetime | None = None,
-    ) -> CanonicalToolCall:
+    ) -> ToolCallRecord:
         """Mark a tool call as failed."""
         resolved_completed_at = completed_at or _utc_now()
         return self._update_call(
@@ -106,7 +139,7 @@ class ToolCallRegistry:
         tool_call_id: ToolCallId,
         reason: str | None = None,
         completed_at: datetime | None = None,
-    ) -> CanonicalToolCall:
+    ) -> ToolCallRecord:
         """Mark a tool call as cancelled."""
         resolved_completed_at = completed_at or _utc_now()
         return self._update_call(
@@ -116,7 +149,7 @@ class ToolCallRegistry:
             completed_at=resolved_completed_at,
         )
 
-    def get(self, tool_call_id: ToolCallId) -> CanonicalToolCall | None:
+    def get(self, tool_call_id: ToolCallId) -> ToolCallRecord | None:
         """Return a tool call by ID."""
         return self._calls.get(tool_call_id)
 
@@ -127,18 +160,18 @@ class ToolCallRegistry:
             raise ValueError(ERROR_TOOL_CALL_NOT_FOUND.format(tool_call_id=tool_call_id))
         return call.args
 
-    def list_calls(self) -> list[CanonicalToolCall]:
+    def list_calls(self) -> list[ToolCallRecord]:
         """Return tool calls in registration order."""
         return [self._calls[tool_call_id] for tool_call_id in self._order]
 
-    def latest_call(self) -> CanonicalToolCall | None:
+    def latest_call(self) -> ToolCallRecord | None:
         """Return the most recently registered tool call."""
         if not self._order:
             return None
         last_tool_call_id = self._order[-1]
         return self._calls.get(last_tool_call_id)
 
-    def recent_calls(self, limit: int) -> list[CanonicalToolCall]:
+    def recent_calls(self, limit: int) -> list[ToolCallRecord]:
         """Return the most recent tool calls."""
         call_ids = self._order[-limit:] if limit > 0 else []
         return [self._calls[tool_call_id] for tool_call_id in call_ids]
@@ -178,7 +211,7 @@ class ToolCallRegistry:
             records.append(record)
         return records
 
-    def _update_call(self, tool_call_id: ToolCallId, **updates: Any) -> CanonicalToolCall:
+    def _update_call(self, tool_call_id: ToolCallId, **updates: Any) -> ToolCallRecord:
         call = self._calls.get(tool_call_id)
         if call is None:
             raise ValueError(ERROR_TOOL_CALL_NOT_FOUND.format(tool_call_id=tool_call_id))
